+++ src/core/CHANGELOG.md
@@ -7,3 +7,16 @@ Implement & Exploration
 - [ ] Create unittest scenarios on `utils`, `engine`, and `io`
 - [ ] Develop mapping models with `postgresql`
 - [ ] Develop mapping models with `csv`, `json`
+
+Tracking
+--------
+
+- [ ] Create config parser class for create instance of configuration instance
+  - [x] Create config parser class
+  - [ ] Create unittest
+    
+- [ ] Create path utility functions 
+  - [ ] Create path utility function
+  - [ ] Create unittest
+    
+- [ ] Create logging function